requirements.txt
pytest
python-dotenv
pylint
pandas